+++ benchmark/qwen3-8b-gptq/quant_benchmark.py
@@ -15,7 +15,7 @@ import itertools
 import hydra
 from omegaconf import DictConfig, OmegaConf
 
-from onecomp import CalibrationConfig, GPTQ, ModelConfig, Runner
+from onecomp import GPTQ, CalibrationConfig, ModelConfig, Runner
 
 
 def create_quantizers(cfg: DictConfig):
@@ -52,7 +52,7 @@ def create_quantizers(cfg: DictConfig):
 def main(cfg: DictConfig):
     print(OmegaConf.to_yaml(cfg))
 
-    model_config = ModelConfig(path=cfg.model_path, device=cfg.model_device)
+    model_config = ModelConfig(path=cfg.model_path, device=cfg.model_device, dtype="bfloat16")
 
     quantizers = create_quantizers(cfg)
 
@@ -75,9 +75,7 @@ def main(cfg: DictConfig):
     runner.run()
 
     for q in quantizers:
-        runner.save_quantization_statistics(
-            f"quantization_statistics_{q.name}.json", quantizer=q
-        )
+        runner.save_quantization_statistics(f"quantization_statistics_{q.name}.json", quantizer=q)
 
     if cfg.calc_ppl:
         runner.benchmark_perplexity(
